fixed build bug
